+++ docs/changelog.rst
@@ -22,6 +22,9 @@ To update |kitty|, :doc:`follow the instructions <binary>`.
 
 - Linux: Workaround for broken Nvidia drivers for old cards (:iss:`456`)
 
+- Wayland: Fix kitty being killed on some Wayland compositors if a hidden window
+  has a lot of output (:iss:`2329`)
+
 
 0.17.4 [2020-05-09]
 --------------------